+++ src/main/java/com/school/system/remark/RemarkService.java
@@ -1,12 +1,21 @@
 package com.school.system.remark;
 
 import com.school.system.exception.NotFoundException;
+import com.school.system.policy.CreatePolicy;
+import com.school.system.policy.DeletePolicy;
+import com.school.system.policy.ReadPolicy;
+import com.school.system.policy.UpdatePolicy;
+import com.school.system.policy.exception.CannotCreateException;
+import com.school.system.policy.exception.CannotDeleteException;
+import com.school.system.policy.exception.CannotReadException;
+import com.school.system.policy.exception.CannotUpdateException;
 import com.school.system.subject.Subject;
 import com.school.system.subject.SubjectRepository;
 import com.school.system.users.student.Student;
 import com.school.system.users.student.StudentRepository;
 import com.school.system.users.teacher.Teacher;
 import com.school.system.users.teacher.TeacherRepository;
+import com.school.system.users.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -35,7 +44,7 @@ public class RemarkService {
         this.subjectRepository = subjectRepository;
     }
 
-    public RemarkResponseDTO createRemark(RemarkRequestDTO remarkDTO) {
+    public RemarkResponseDTO createRemark(RemarkRequestDTO remarkDTO, User user) {
         Remark toCreate = new Remark();
 
         Teacher teacher = teacherRepository.findById(remarkDTO.teacher())
@@ -52,6 +61,10 @@ public class RemarkService {
         toCreate.setHeading(remarkDTO.heading());
         toCreate.setBody(remarkDTO.body());
 
+        if(!CreatePolicy.canCreateRemark(user, toCreate)) {
+            throw new CannotCreateException();
+        }
+
         return RemarkMapper.remarkToRemarkResponseDTO(remarkRepository.save(toCreate));
     }
 
@@ -61,9 +74,24 @@ public class RemarkService {
         return remarks.map(remark -> RemarkMapper.remarkToRemarkResponseDTO(remark));
     }
 
-    public RemarkResponseDTO updateRemark(UUID id, RemarkRequestDTO remarkDTO) {
+    public RemarkResponseDTO getRemark(UUID id, User user) {
+        Remark remark = remarkRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Remark not found"));
+
+        if(!ReadPolicy.canReadRemark(user, remark)) {
+            throw new CannotReadException();
+        }
+
+        return RemarkMapper.remarkToRemarkResponseDTO(remark);
+    }
+
+    public RemarkResponseDTO updateRemark(UUID id, RemarkRequestDTO remarkDTO, User user) {
         Remark toUpdate = remarkRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Remark not found"));
+
+        if(!UpdatePolicy.canUpdateRemark(user, toUpdate)) {
+            throw new CannotUpdateException();
+        }
 
         if(toUpdate.getTeacher().getId() != remarkDTO.teacher()) {
             Teacher teacher = teacherRepository.findById(remarkDTO.teacher())
@@ -89,9 +117,14 @@ public class RemarkService {
         return RemarkMapper.remarkToRemarkResponseDTO(remarkRepository.save(toUpdate));
     }
 
-    public void deleteRemark(UUID id) {
+    public void deleteRemark(UUID id, User user) {
         Remark toDelete = remarkRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Remark not found"));
+
+        if(!DeletePolicy.canDeleteRemark(user, toDelete)) {
+            throw new CannotDeleteException();
+        }
+
         remarkRepository.delete(toDelete);
     }
 }
