+++ src/main/java/com/school/system/subject/SubjectService.java
@@ -3,8 +3,17 @@ package com.school.system.subject;
 import com.school.system.exception.NotFoundException;
 import com.school.system.grade.Grade;
 import com.school.system.grade.GradeRepository;
+import com.school.system.policy.CreatePolicy;
+import com.school.system.policy.DeletePolicy;
+import com.school.system.policy.ReadPolicy;
+import com.school.system.policy.UpdatePolicy;
+import com.school.system.policy.exception.CannotCreateException;
+import com.school.system.policy.exception.CannotDeleteException;
+import com.school.system.policy.exception.CannotReadException;
+import com.school.system.policy.exception.CannotUpdateException;
 import com.school.system.users.teacher.Teacher;
 import com.school.system.users.teacher.TeacherRepository;
+import com.school.system.users.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -29,7 +38,7 @@ public class SubjectService {
         this.gradeRepository = gradeRepository;
     }
 
-    public SubjectResponseDTO createSubject(SubjectRequestDTO subjectDTO) {
+    public SubjectResponseDTO createSubject(SubjectRequestDTO subjectDTO, User user) {
         Subject toCreate = new Subject();
 
         Teacher teacher = teacherRepository.findById(subjectDTO.teacher())
@@ -43,6 +52,10 @@ public class SubjectService {
         toCreate.setTeacher(teacher);
         toCreate.setSchoolClass(grade);
 
+        if(!CreatePolicy.canCreateSubject(user, toCreate)) {
+            throw new CannotCreateException();
+        }
+
         return SubjectMapper.subjectToSubjectResponseDTO(subjectRepository.save(toCreate));
     }
 
@@ -52,8 +65,23 @@ public class SubjectService {
         return subjects.map(subject -> SubjectMapper.subjectToSubjectResponseDTO(subject));
     }
 
-    public SubjectResponseDTO updateSubject(UUID id, SubjectRequestDTO subjectDTO) {
+    public SubjectResponseDTO getSubject(UUID id, User user) {
+        Subject subject = subjectRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Subject not found"));
+
+        if(!ReadPolicy.canReadSubject(user, subject)) {
+            throw new CannotReadException();
+        }
+
+        return SubjectMapper.subjectToSubjectResponseDTO(subject);
+    }
+
+    public SubjectResponseDTO updateSubject(UUID id, SubjectRequestDTO subjectDTO, User user) {
         Subject toUpdate = subjectRepository.findById(id).orElseThrow(() -> new NotFoundException("Subject not found"));
+
+        if(!UpdatePolicy.canUpdateSubject(user, toUpdate)) {
+            throw new CannotUpdateException();
+        }
 
         toUpdate.setName(subjectDTO.name());
         toUpdate.setSemester(subjectDTO.semester());
@@ -73,9 +101,14 @@ public class SubjectService {
         return SubjectMapper.subjectToSubjectResponseDTO(subjectRepository.save(toUpdate));
     }
 
-    public void deleteSubject(UUID id) {
+    public void deleteSubject(UUID id, User user) {
         Subject toDelete = subjectRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Subject not found"));
+
+        if(!DeletePolicy.canDeleteSubject(user, toDelete)) {
+            throw new CannotDeleteException();
+        }
+
         subjectRepository.delete(toDelete);
     }
 }
